tfdbg: Some changes that take effect only internally. Change: 141311719
